+++ scctool/settings/logoManager.py
@@ -6,9 +6,9 @@ import logging
 import os
 import shutil
 from time import time
-from urllib.request import urlopen, urlretrieve
 
 import humanize
+import requests
 from PyQt5.QtCore import Qt
 from PyQt5.QtGui import QPixmap
 
@@ -389,7 +389,7 @@ class Logo:
 
         return True
 
-    def fromURL(self, url, download=True, localFile=None):
+    def fromURL(self, url, download=True, localFile=None, verify=True):
         _, ext = os.path.splitext(url)
         self._format = ext.split("?")[0].replace(".", "").lower()
         self.generateIdentifier()
@@ -399,23 +399,28 @@ class Logo:
         if not download:
             return fname
 
+        needs_download = False
         if localFile:
             with open(localFile, "rb") as in_file:
-                local_byte = in_file.read(512)
+                local_chunk = in_file.read(512)
 
-            file = urlopen(url)
-            data = file.read(512)
-            if(data == local_byte):
-                needs_download = False
-            else:
-                needs_download = True
+            r = requests.get(url, stream=True, verify=verify)
+            if r.headers.get('Content-Type', 'text').find('text') != -1:
+                module_logger.info('Logo is no valid image.')
+                return False
+            if r.status_code == 200:
+                with open(fname, 'wb') as f:
+                    for chunk in r.iter_content(512):
+                        if not needs_download and chunk == local_chunk:
+                            break
+                        else:
+                            needs_download = True
+                        f.write(chunk)
 
         if needs_download:
-            urlretrieve(url, fname)
             self.refreshData()
-            return True
-        else:
-            return False
+
+        return needs_download
 
     def refreshData(self):
         file = self.getAbsFile()